+++ CHANGES.rst
@@ -5,7 +5,10 @@ Changes
 2.3.0
 =====
 
-TBD
+Application Changes
+-------------------
+
+* Add support for decimal column and values for panelist Lightning round start and correct
 
 2.2.0
 =====